Migrated to new init logic, implemented Peers route
Highway requires new init logic to be able to access all currently connected peers.
After a little bit of reverse engineering deemed this to be the least hacky way to obtain access.

Initial version of Peers. This route can be used to retrieve client ids, which are required for piping.