Migration to Highway, incorporate piping, hostname changing
The port of the Wallaby client has kicked off and with it come lots of broken features.
WallabyControl will have to be remade to incorporate piping and routes for sensor readouts as well as wi-fi network selection have to be added.

As a quick piping demo the Hostname changer route was implemented which already works in the Sublime Text client.